adjust ExceededSigOpLimit to u16 (fixes overflow)

+++ crypto/txscript/src/runtime_sig_op_counter.rs
@@ -46,14 +46,13 @@ impl RuntimeSigOpCounter {
     /// counter.consume_sig_op().unwrap_err(); // Err(ExceededSigOpLimit)
     /// ```
     pub fn consume_sig_op(&mut self) -> Result<(), TxScriptError> {
-        self.sig_op_remaining =
-            self.sig_op_remaining.checked_sub(1).ok_or(TxScriptError::ExceededSigOpLimit(self.sig_op_limit as u8))?;
+        self.sig_op_remaining = self.sig_op_remaining.checked_sub(1).ok_or(TxScriptError::ExceededSigOpLimit(self.sig_op_limit))?;
         Ok(())
     }
 
     pub fn consume_sig_ops(&mut self, count: u16) -> Result<(), TxScriptError> {
         self.sig_op_remaining =
-            self.sig_op_remaining.checked_sub(count).ok_or(TxScriptError::ExceededSigOpLimit(self.sig_op_limit as u8))?;
+            self.sig_op_remaining.checked_sub(count).ok_or(TxScriptError::ExceededSigOpLimit(self.sig_op_limit))?;
         Ok(())
     }
